service-bus/requirements.txt
azure-servicebus==7.8.0
python-dotenv==0.21.0